Move unit tests from type-inference to match-inference
